microservices/shared-libs/requirements.txt
requests==2.31.0
pika==1.3.2
pyjwt==2.8.0
redis==5.0.1 